feat: Implement simulation to place orders to stores

+++ central/src/main/resources/db/migration/V2_3_0__create_store_revenue.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS store_revenue (
     store_id BIGINT PRIMARY KEY,
-    total DECIMAL NOT NULL,
+    total FLOAT NOT NULL,
     CONSTRAINT fk_store FOREIGN KEY(store_id) REFERENCES stores(id)
 );
